# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MCM Re:SENSE API
  version: 1.0.0
  description: Customer Web, Advisor Web, Store Reveal을 sessionId로 연결하는 해커톤 MVP API
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /sessions:
    post:
      tags: [Customer]
      summary: 경험 세션 생성
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Customer]
      summary: 전체 세션 조회
      operationId: getSession
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /sessions/{sessionId}/preferences:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    put:
      tags: [Customer]
      summary: 취향 저장 또는 덮어쓰기
      description: 기존 취향을 변경하면 하위 Intent와 UNSEEN 결과가 초기화됩니다.
      operationId: savePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceRequest'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{sessionId}/intent:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [AI]
      summary: Intent Profile 생성
      operationId: interpretIntent
      responses:
        '200':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidState'
  /sessions/{sessionId}/unseen:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [AI]
      summary: 비동기 UNSEEN 생성 요청
      operationId: requestUnseen
      responses:
        '202':
          description: 생성 요청 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnseenResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{sessionId}/unseen/selection:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    patch:
      tags: [AI]
      summary: 피그마 결과 카드에서 최종 UNSEEN 후보 선택
      operationId: selectUnseenCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectUnseenCandidateRequest'
      responses:
        '200':
          description: 선택 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnseenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidState'
        '422':
          $ref: '#/components/responses/InvalidState'
    get:
      tags: [AI]
      summary: UNSEEN 생성 상태 조회
      operationId: getUnseen
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnseenResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /stores:
    get:
      tags: [Reservation]
      summary: 예약 가능 매장 조회
      operationId: getStores
      parameters:
        - name: city
          in: query
          required: false
          description: 위치 권한 또는 직접 선택으로 얻은 도시명 필터
          schema:
            type: string
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoreResponse'
  /stores/{storeId}/slots:
    get:
      tags: [Reservation]
      summary: 날짜별 예약 슬롯 조회
      operationId: getSlots
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SlotResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations:
    post:
      tags: [Reservation]
      summary: 매장 경험 예약
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: 예약 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/InvalidState'
  /reservations/{reservationId}:
    parameters:
      - name: reservationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Reservation]
      summary: 방문 전 예약 변경
      operationId: updateReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationUpdateRequest'
      responses:
        '200':
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
    delete:
      tags: [Reservation]
      summary: 방문 전 예약 취소
      operationId: cancelReservation
      responses:
        '204':
          description: 취소 성공
  /reservations/{reservationId}/calendar.ics:
    parameters:
      - name: reservationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Reservation]
      summary: 예약 확인 화면의 캘린더 파일 다운로드
      operationId: downloadReservationCalendar
      responses:
        '200':
          description: iCalendar 파일
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="resense-770e8400-e29b-41d4-a716-446655440000.ics"
          content:
            text/calendar:
              schema:
                type: string
                example: "BEGIN:VCALENDAR\r\nVERSION:2.0\r\n..."
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{sessionId}/reservation:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Reservation]
      summary: 세션 예약 조회
      operationId: getReservation
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /sessions/{sessionId}/feedback:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Feedback]
      summary: 구매 결과와 Loved/Concern/Wants 저장
      operationId: saveFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /customers/{demoCustomerId}/memory:
    get:
      tags: [Feedback]
      summary: 데모 고객의 과거 경험 조회
      operationId: getCustomerMemory
      parameters:
        - name: demoCustomerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 최신순 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerMemoryResponse'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 요청값 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 예약 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidState:
      description: 현재 Journey 상태에서 처리할 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ExperienceStatus:
      type: string
      enum: [CREATED, PREFERENCES_SAVED, INTENT_READY, UNSEEN_PROCESSING, UNSEEN_READY, RESERVED, ARRIVED, PERSONAL_EDIT_READY, REVEALING, COMPLETED]
    GenerationStatus:
      type: string
      enum: [NOT_STARTED, PROCESSING, READY, FAILED]
    ReservationStatus:
      type: string
      enum: [BOOKED, ARRIVED, COMPLETED, CANCELLED]
    RevealStage:
      type: string
      enum: [NOT_STARTED, WELCOME, UNSEEN_REVEAL, LIFESTYLE_SCENE, FINAL_TRANSITION, COMPLETED]
    PurchaseResult:
      type: string
      enum: [PURCHASED, NOT_PURCHASED, UNDECIDED]
...
    CreateSessionRequest:
      type: object
      required: [demoCustomerId, customerName, dataConsent]
      properties:
        demoCustomerId:
          type: string
          maxLength: 80
          example: demo-lena
        customerName:
          type: string
          maxLength: 80
          example: Lena
        phone:
          type: string
          maxLength: 30
          nullable: true
          example: 010-1234-5678
        email:
          type: string
          format: email
          maxLength: 160
          nullable: true
          example: lena@example.com
        gender:
          type: string
          maxLength: 30
          nullable: true
          example: FEMALE
        dataConsent:
          type: boolean
          example: true
    PreferenceRequest:
      type: object
      required: [silhouette, structure, proportion, color, attitude, contexts, lockedAttribute]
      properties:
        silhouette:
          type: string
          example: Crossbody
        structure:
          type: string
          example: Soft
        proportion:
          type: string
          example: Compact-Spacious
        color:
          type: string
          example: Cognac
        attitude:
          type: string
          example: Quiet
        contexts:
          type: array
          minItems: 1
          items:
            type: string
          example: [Work, Travel]
        lockedAttribute:
          type: string
          enum: [Shape, Color, Space, Attitude]
          example: Shape
    PreferenceResponse:
      allOf:
        - $ref: '#/components/schemas/PreferenceRequest'
...
    IntentProfileResponse:
      type: object
      properties:
        purpose:
          type: string
          example: Work + Travel
        priority:
          type: string
          example: Mobility / Lightweight
        style:
          type: string
          example: Relaxed / Modern
        signature:
          type: string
          example: Soft Structure / Cognac
        lockedAttribute:
          type: string
          example: Shape
        concern:
          type: string
          example: Styling Versatility
        summary:
          type: string
    UnseenResponse:
      type: object
      properties:
        unseenId:
          type: string
          nullable: true
          example: UNSEEN-50256B70
        status:
          $ref: '#/components/schemas/GenerationStatus'
        imageUrl:
          type: string
          nullable: true
          example: /assets/unseen/image4.png
        error:
          type: string
          nullable: true
        candidates:
          type: array
          description: READY일 때 피그마 결과 카드에 표시할 최대 4개 후보
          items:
            $ref: '#/components/schemas/UnseenCandidateResponse'
        selectedCandidateId:
          type: string
          format: uuid
          nullable: true
    UnseenCandidateResponse:
      type: object
      properties:
        candidateId:
          type: string
          format: uuid
        imageUrl:
          type: string
          example: /assets/unseen/image4.png
        shape:
          type: string
          example: Top Handle
        size:
          type: string
          example: Mini
        color:
          type: string
          example: Cognac
        rank:
          type: integer
          example: 1
        selected:
          type: boolean
          example: false
    SelectUnseenCandidateRequest:
      type: object
      required: [candidateId]
      properties:
        candidateId:
          type: string
          format: uuid
    SessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        demoCustomerId:
          type: string
        customerName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        gender:
          type: string
          nullable: true
        dataConsent:
          type: boolean
        status:
          $ref: '#/components/schemas/ExperienceStatus'
        preferences:
          allOf:
            - $ref: '#/components/schemas/PreferenceResponse'
          nullable: true
        intentProfile:
          allOf:
            - $ref: '#/components/schemas/IntentProfileResponse'
          nullable: true
        unseen:
          $ref: '#/components/schemas/UnseenResponse'
        advisorPriority:
          type: string
          nullable: true
        revealStage:
          $ref: '#/components/schemas/RevealStage'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StoreResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        address:
          type: string
    SlotResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        scheduledAt:
          type: string
          format: date-time
        available:
          type: boolean
    ReservationRequest:
      type: object
      required: [sessionId, storeId, scheduledAt]
      properties:
        sessionId:
          type: string
          format: uuid
        storeId:
          type: integer
          format: int64
        scheduledAt:
          type: string
          format: date-time
    ReservationUpdateRequest:
      type: object
      required: [storeId, scheduledAt]
      properties:
        storeId:
          type: integer
          format: int64
        scheduledAt:
          type: string
          format: date-time
    ReservationResponse:
      type: object
      properties:
        reservationId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        storeId:
          type: integer
          format: int64
        storeName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ReservationStatus'
        passCode:
          type: string
          example: PASS-82D1FC80
...
    FeedbackRequest:
      type: object
      required: [result]
      properties:
        loved:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        wants:
          type: array
          items:
            type: string
        result:
          $ref: '#/components/schemas/PurchaseResult'
        purchasedProductId:
          type: integer
          format: int64
          nullable: true
    FeedbackResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        result:
          $ref: '#/components/schemas/PurchaseResult'
        status:
          $ref: '#/components/schemas/ExperienceStatus'
    MemoryItem:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        unseenId:
          type: string
        intentSummary:
          type: string
        loved:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        wants:
          type: array
          items:
            type: string
        result:
          $ref: '#/components/schemas/PurchaseResult'
        experiencedAt:
          type: string
          format: date-time
    CustomerMemoryResponse:
      type: object
      properties:
        demoCustomerId:
          type: string
        experiences:
          type: array
          items:
            $ref: '#/components/schemas/MemoryItem'
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string